test::proptest::with_binary_with_atom_without_name_encoding_errors_badarg
