+++ lumen_runtime/src/test/proptest.rs
@@ -264,6 +264,31 @@ pub fn with_binary_without_atom_encoding_errors_badarg(
     );
 }
 
+pub fn with_binary_with_atom_without_name_encoding_errors_badarg(
+    source_file: &'static str,
+    native: fn(Term, Term) -> exception::Result<Term>,
+) {
+    run(
+        source_file,
+        |arc_process| {
+            (
+                super::strategy::term::is_binary(arc_process.clone()),
+                super::strategy::term::atom::is_not_encoding(),
+            )
+        },
+        |(binary, encoding)| {
+            let encoding_atom: Atom = encoding.try_into().unwrap();
+
+            prop_assert_badarg!(
+                        native(binary, encoding),
+                        format!("invalid atom encoding name: '{0}' is not one of the supported values (latin1, unicode, or utf8)", encoding_atom.name())
+                    );
+
+            Ok(())
+        },
+    );
+}
+
 pub fn with_integer_left_without_integer_right_errors_badarith(
     source_file: &'static str,
     native: fn(&Process, Term, Term) -> exception::Result<Term>,
